% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{alexaResponse}
\alias{alexaResponse}
\title{Prepare a Response for Alexa}
\usage{
alexaResponse(output, card, reprompt, endSession = FALSE)
}
\description{
Form a response object that satisfies the requirements of a response to an
Alexa custom skill request.
}
