% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{dispatchAlexaRequest}
\alias{dispatchAlexaRequest}
\title{Dispatch Alexa Request to Handlers}
\usage{
dispatchAlexaRequest(req, intent = function(name, slots, ...) { },
  launch = function(...) { }, end = function(...) { })
}
\arguments{
\item{req}{The plumber request to dispatch on.}

\item{intent}{The function to invoke if the request is of type \code{IntentRequest}.}

\item{launch}{The function to invoke if the request is of type \code{LaunchRequest}.}

\item{end}{The function to invoke if the request is of type \code{SessionEndedRequest}.}
}
\description{
A convenience function that routes an incoming plumber request to the
appropriate handler given the request's type.
}
\details{
Any of the three parameters that are not provided will just function as no-ops.

See
https://developer.amazon.com/public/solutions/alexa/alexa-skills-kit/docs/custom-standard-request-types-reference
for complete details on the three intent types.

All provided handlers should accept \code{...} as this API is subject to
change and new arguments may be added. Having your handler functions accept
\code{...} is a way of ensuring that they won't break if/when alexar begins
passing in new parameters that your code doesn't recognize or use.
}
