% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validateAlexaRequest}
\alias{validateAlexaRequest}
\title{Validate an Alexa Request}
\usage{
validateAlexaRequest(req)
}
\arguments{
\item{req}{The incoming plumber request to validate}
}
\description{
See https://developer.amazon.com/public/solutions/alexa/alexa-skills-kit/docs/developing-an-alexa-skill-as-a-web-service#h2_verify_sig_cert for more details.
}
