% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{alexaResponse}
\alias{alexaResponse}
\title{Prepare a Response for Alexa}
\usage{
alexaResponse(output, card, reprompt, sessionAttributes = NULL,
  endSession = FALSE)
}
\arguments{
\item{output}{A character string containing the words Alexa should read back
to the request. Converted to \code{PlainTextSpeech} in the Alexa API.}

\item{card}{A list containing the contents of the card to be displayed in
response to the request. See
https://developer.amazon.com/public/solutions/alexa/alexa-skills-kit/docs/alexa-skills-kit-interface-reference#card-object
for more information about how the card should be structured.}

\item{reprompt}{A character string representing the \code{PlainTextSpeech} to be
included as the reprompt object. From the Amazon documentation: "The object
containing the outputSpeech to use if a re-prompt is necessary. This is
used if the your service keeps the session open after sending the response,
but the user does not respond with anything that maps to an intent defined
in your voice interface while the audio stream is open. If this is not set,
the user is not re-prompted."
https://developer.amazon.com/public/solutions/alexa/alexa-skills-kit/docs/alexa-skills-kit-interface-reference#response-object}

\item{sessionAttributes}{A list of data to attach to this Alexa session.
Subsequent requests from this session will include whatever data you
provide here in the \code{attributes} field.}

\item{endSession}{If \code{TRUE}, will force the session to close as a result of
this response.}
}
\description{
Form a response object that satisfies the requirements of a response to an
Alexa custom skill request. All objects returned for an Alexa custom skill
should be an object formed using this function.
}
