% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{dispatchAlexaRequest}
\alias{dispatchAlexaRequest}
\title{Dispatch Alexa Request to Handlers}
\usage{
dispatchAlexaRequest(req, intent = function(name, slots, ...) { },
  launch = function(...) { }, end = function(...) { })
}
\description{
Dispatch Alexa Request to Handlers
}
