% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validateAlexaRequest}
\alias{validateAlexaRequest}
\title{Validate an Alexa Request}
\usage{
validateAlexaRequest(req)
}
\arguments{
\item{req}{The incoming plumber request to validate}
}
\description{
Validate that an incoming request is legitimate (i.e. from Amazon). This
validation on incoming requests is not only important for security, but also
required in order for Amazon to certify your Alexa custom skill.
}
\details{
See
https://developer.amazon.com/public/solutions/alexa/alexa-skills-kit/docs/developing-an-alexa-skill-as-a-web-service#h2_verify_sig_cert
for more details.
}
